+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 ## Unreleased
 
+## 1.21.0
+
+- **New**
+  - Added support for `Dragon Turret` (🎨 by @Chasarooni) (✋ @tunderpower)
+- **Updated**
+  - Handled situation where a token has no actor breaking `Wall of Fire` (🐛 @zionhian)
+  - Added `Floating Flame` and `Avenging Wildwood` to summon's who's DCs are linked to their summoner
+  - Updated chinese translation (🌐 @AlphaStarGuide)
+
 ## 1.20.3
 
 - Push data from `0.20.2`